requirements.txt
aiocron==latest
aiohttp==latest
aiosignal==latest
anyio==latest
async-timeout==latest
attrs==latest
beautifulsoup4==latest
certifi==latest
charset-normalizer==latest
croniter==latest
discord.py==latest
exceptiongroup==latest
frozenlist==latest
h11==latest
httpcore==latest
httpx==latest
idna==latest
multidict==latest
pip==latest
python-dateutil==latest
python-dotenv==latest
pytz==latest
requests==latest
setuptools==latest
six==latest
sniffio==latest
soupsieve==latest
tzdata==latest
tzlocal==latest
urllib3==latest
wheel==latest
yarl==latest
